feat: move crawling logic into workers
