+++ packages/backend/src/crawling/workerWrapper.js
@@ -1,0 +1,6 @@
+const { workerData } = require('worker_threads');
+
+if (workerData.fullpath.endsWith(".ts")) {
+    require('ts-node').register();
+}
+module.exports = require(workerData.fullpath);
